datapath/linux: drop dependency on pkg/option
With all the preceding commits, pkg/datapath/linux/route is now the only
package imported in the cilium-cni and cilium-docker plugins which pulls
in pkg/option. This can easily be avoided by changing
NodeDeviceWithDefaultRoute to not fetch the EnableIPv{4,6} options
itself from pkg/option, but by taking them as arguments.

Together with the preceding commits, this reduces the binary size of
cilium-cni and cilium-docker by ~1 MiB each:

before:

-rwxr-xr-x 1 tklauser tklauser 15364096 Mai  4 16:39 plugins/cilium-cni/cilium-cni
-rwxr-xr-x 1 tklauser tklauser 17399808 Mai  4 16:39 plugins/cilium-docker/cilium-docker

after:

-rwxr-xr-x 1 tklauser tklauser 13938688 Mai  4 17:00 plugins/cilium-cni/cilium-cni
-rwxr-xr-x 1 tklauser tklauser 16523264 Mai  4 17:00 plugins/cilium-docker/cilium-docker

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>